Add InfestedStoneDecorator
